reorganize whiskey table components. create lazyQuery for search (WIP)
